WTA #44: Moved files around.
