+++ home/src/main/java/com/jacob/wakatimeapp/home/domain/Utils.kt
@@ -1,0 +1,9 @@
+package com.jacob.wakatimeapp.home.domain
+
+import com.jacob.wakatimeapp.core.models.Time
+import kotlinx.datetime.LocalDate
+
+fun Map<LocalDate, Time>.getLatestStreakInRange() = toSortedMap()
+    .entries
+    .reversed()
+    .takeWhile { it.value != Time.ZERO }
